+++ subprojects/lua-cjson/scripts/lua/json2lua.sh
@@ -1,2 +1,2 @@
 BASEDIR=$(dirname $0)
-exec "$BASEDIR/lua" -e "json=require('cjson');util=require('cjson.util');json_text=util.file_load(arg[1]);t=json.decode(json_text);print(util.serialise_value(t));" "$@"
+exec "$BASEDIR/lua" "$BASEDIR/../lib/luarocks/rocks-5.4/lua-cjson/2.1.0.10-1/bin/json2lua.lua" "$@"